libflux/attr: pull in flux_get_rank(), _size()
Problem: flux_get_rank() and flux_get_size() are just
wrappers around flux_attr_get() and don't warrant having
their own header and c module.

Incorporate into attr.h and attr.c.

The remaining bits of info.[ch] will be dealt with in
a subsequent commit.

Update users.
