+++ src/common/libflux/attr.c
@@ -177,6 +177,26 @@ int flux_attr_set_cacheonly (flux_t *h, const char *name, const char *val)
     return 0;
 }
 
+int flux_get_size (flux_t *h, uint32_t *size)
+{
+    const char *val;
+
+    if (!(val = flux_attr_get (h, "size")))
+        return -1;
+    *size = strtoul (val, NULL, 10);
+    return 0;
+}
+
+int flux_get_rank (flux_t *h, uint32_t *rank)
+{
+    const char *val;
+
+    if (!(val = flux_attr_get (h, "rank")))
+        return -1;
+    *rank = strtoul (val, NULL, 10);
+    return 0;
+}
+
 /*
  * vi:tabstop=4 shiftwidth=4 expandtab
  */
